Feature/still in use

+++ src/main/java/be/tribersoft/sensor/domain/impl/type/TypeFacadeImpl.java
@@ -3,8 +3,10 @@ package be.tribersoft.sensor.domain.impl.type;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import be.tribersoft.sensor.domain.api.sensor.SensorRepository;
 import be.tribersoft.sensor.domain.api.type.TypeFacade;
 import be.tribersoft.sensor.domain.api.type.TypeMessage;
+import be.tribersoft.sensor.domain.api.type.exception.TypeStillInUseException;
 
 @Named
 public class TypeFacadeImpl implements TypeFacade {
@@ -15,6 +17,8 @@ public class TypeFacadeImpl implements TypeFacade {
 	private TypeFactory typeFactory;
 	@Inject
 	private TypeUpdater typeUpdater;
+	@Inject
+	private SensorRepository sensorRepository;
 
 	@Override
 	public void save(TypeMessage typeMessage) {
@@ -29,6 +33,9 @@ public class TypeFacadeImpl implements TypeFacade {
 
 	@Override
 	public void delete(String id, Long version) {
+		if (sensorRepository.typeInUse(id)) {
+			throw new TypeStillInUseException();
+		}
 		TypeEntity type = typeRepository.getByIdAndVersion(id, version);
 		typeRepository.delete(type);
 	}
